Change model_no datatypes - Fix issue on storing random number in database - Change model_no datatype from Integer to BigInt

+++ src/models/courier.js
@@ -38,7 +38,7 @@ module.exports = (sequelize, DataTypes) => {
 				defaultValue: DataTypes.UUIDV4,
 			},
 			courier_no: {
-				type: DataTypes.INTEGER(11),
+				type: DataTypes.BIGINT(11),
 				allowNull: false,
 				unique: { msg: 'Courier Number already exists.' },
 				validate: {
