+++ src/models/deliverydetails.js
@@ -42,7 +42,7 @@ module.exports = (sequelize, DataTypes) => {
 				defaultValue: DataTypes.UUIDV4,
 			},
 			tracking_no: {
-				type: DataTypes.INTEGER(11),
+				type: DataTypes.BIGINT(11),
 				allowNull: false,
 				unique: { msg: 'Tracking Number already exists.' },
 				validate: {
